requirements.txt
pyTelegramBotAPI
google-generativeai
python-dotenv
rich
sqlalchemy
pypdf
lancedb
requests
plotly
reportlab
streamlit
pandas